Editor: fixed alignment of categories/tags info-popover icons
Previously the info icon for tags was too close to the token field and the category selector had different spacing. Both of their labels were misaligned with their respective info-icons. I increased the line-height of the labels to make room for the icon and removed excess top margin from the category editor. A clearfix was also reqired since the icons are floated right.